change commas to locale
