+++ lib/Map/LegendHelper.js
@@ -7,7 +7,7 @@ var defined = require('terriajs-cesium/Source/Core/defined');
 var definedNotNull = require('terriajs-cesium/Source/Core/definedNotNull');
 
 var ColorMap = require('../Map/ColorMap');
-var formatNumberWithCommas = require('../Core/formatNumberWithCommas');
+var formatNumberToLocale = require('../Core/formatNumberToLocale');
 var Legend = require('../Map/Legend');
 var TableStyle = require('../Models/TableStyle');
 var VarType = require('../Map/VarType');
@@ -297,7 +297,7 @@ function buildBinColors(legendHelper, colorBins) {
 
 
 function convertToStringWithAtMostTwoDecimalPlaces(f) {
-    return formatNumberWithCommas(Math.round(f * 100) / 100, 5);  // Use commas once there are over 5 digits before the decimal point.
+    return formatNumberToLocale(Math.round(f * 100) / 100, 5);  // Use commas once there are over 5 digits before the decimal point.
 }
 
 function convertColorArrayToCssString(colorArray) {
