+++ lib/ViewModels/FeatureInfoPanelSectionViewModel.js
@@ -9,7 +9,7 @@ var destroyObject = require('terriajs-cesium/Source/Core/destroyObject');
 var isArray = require('terriajs-cesium/Source/Core/isArray');
 var FeatureDetection = require('terriajs-cesium/Source/Core/FeatureDetection');
 
-var formatNumberWithCommas = require('../Core/formatNumberWithCommas');
+var formatNumberToLocale = require('../Core/formatNumberToLocale');
 var loadView = require('../Core/loadView');
 var TerriaError = require('../Core/TerriaError');
 var svgArrowDown = require('../SvgPaths/svgArrowDown');
@@ -23,7 +23,7 @@ var CAN_USE_DATA_URI_IN_HREF = !(FeatureDetection.isInternetExplorer() || /Edge/
 
 // A Mustache function used to insert commas into numbers
 function addCommas(text, render) {
-    return formatNumberWithCommas(render(text));
+    return formatNumberToLocale(render(text));
 }
 
 /**
@@ -301,7 +301,7 @@ function propertyGetTimeValues(properties, clock) {
 //     for (var key in properties) {
 //         if (properties.hasOwnProperty(key)) {
 //             if (typeof properties[key] === 'number') {
-//                 properties[key] = formatNumberWithCommas(properties[key]);
+//                 properties[key] = formatNumberToLocale(properties[key]);
 //             }
 //         }
 //     }
